refactor(routers): move fixed routers to separate package
