+++ src/main/java/ru/zgz/star/backend/routes/AccountRouter.java
@@ -1,8 +1,10 @@
-package ru.zgz.star.backend.routers;
+package ru.zgz.star.backend.routes;
 
 import com.google.gson.Gson;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import ru.zgz.star.backend.daos.AccountDao;
+import ru.zgz.star.backend.models.Account;
 import spark.Request;
 import spark.Response;
 
@@ -13,12 +15,9 @@ public class AccountRouter {
 
   public static String getExactAccount(Request request, Response response) {
     response.type("application/json");
-//    List<Account> accounts = new QAccount().findList();
-//    DAO<Account> dao = new DAO<>(Account.class);
-//    String result = dao.findById(request.params("id"));
-//    logger.info("Account {} found", result);
-//    return result;
-    return "";
+    AccountDao dao = new AccountDao();
+    Account account = dao.getById(request.params("id"));
+    return new Gson().toJson(account);
   }
 
   public static String deleteExactAccount(Request request, Response response) {
